==> EoI_Process/analysis_notebooks/01-eda-and-profile.ipynb <==
import os
import pandas as pd
from ydata_profiling import ProfileReport

unique_merged_data = pd.read_csv("../EoIs/Outputs/Combined/Inputs/unique_merged_data.csv")

profile = ProfileReport(unique_merged_data, title="Profiling Report")
# When you want to share the profile report with external partners, you can use the sensitive mode
#profile = ProfileReport(merged_data, title="Profiling Report", sensitive=True)
folder_path = "../EoIs/Outputs/Reports/"
# create the output folder if it does not exist
os.makedirs(os.path.dirname(folder_path), exist_ok=True)
profile.to_file(folder_path  + "unique_merged_data_report.html")

# Get all the data that "Other (1)" column is not "No"
other_industries = unique_merged_data[unique_merged_data["Other (1)"] != "No"]
other_industries.head(2)

other_industries["Other (1)"].head(2)

other_industries.shape[0]

# save it to a txt file
with open("../EoIs/Outputs/Combined/Outputs/other_industries.txt", "w") as f:
    f.write("\n".join(other_industries["Other (1)"].tolist()))

unique_merged_data["What is the business challenge that your organisation would most like to solve with data science and/or AI? (max 200 words)"].head(2)

with open("../EoIs/Outputs/Combined/Outputs/business_challenges.txt", "w") as f:
    f.write("\n".join(unique_merged_data["What is the business challenge that your organisation would most like to solve with data science and/or AI? (max 200 words)"].tolist()))

# get the count of Agriculture == "Yes" and then the other sectors
agriculture = unique_merged_data[unique_merged_data["Agriculture and food processing"] == "Yes"]
agriculture_count = agriculture.shape[0]

creative = unique_merged_data[unique_merged_data["Creative Industries"] == "Yes"]
creative_count = creative.shape[0]

construction = unique_merged_data[unique_merged_data["Construction"] == "Yes"]
construction_count = construction.shape[0]

transportation = unique_merged_data[unique_merged_data["Transportation, including logistics and warehousing"] == "Yes"]
transportation_count = transportation.shape[0]

# print the counts
print("Agriculture and food processing count:", agriculture_count)
print("Creative industries count:", creative_count)
print("Construction count:", construction_count)
print("Transportation, including logistics and warehousing count:", transportation_count)
