==> EoI_Process/analysis_notebooks/02-sectoral-analysis.ipynb <==
import pandas as pd

data = pd.read_csv("../EoIs/Outputs/Combined/Inputs/unique_merged_data.csv")

# Count the "Yes" and "No" of Data scientists, Machine learning experts, Software engineers, Data analysts, Statisticians columns
data_scientists = data["Data scientists"].value_counts()
print(data_scientists)
machine_learning_experts = data["Machine learning experts"].value_counts()
print(machine_learning_experts)
software_engineers = data["Software engineers"].value_counts()
print(software_engineers)
data_analysts = data["Data analysts"].value_counts()
print(data_analysts)
statisticians = data["Statisticians"].value_counts()
print(statisticians)

print("Count of all businesses that have all 5 roles: ",
    data[(data["Data scientists"] == "Yes") & 
        (data["Machine learning experts"] == "Yes") & 
        (data["Software engineers"] == "Yes") & 
        (data["Data analysts"] == "Yes") & 
        (data["Statisticians"] == "Yes")].shape[0])

# Now analyse these numbers for Agriculture and food processing, Construction, Creative Industries, and Transportation, including logistics and warehousing sectors, and Other. Each sector is represented as a column in the data frame. These columns contain values of "Yes" and "No" for each profession. Analyse the count of professionals in each sector.

# get the rows where the sector is Agriculture and food processing
agriculture = data[data["Agriculture and food processing"] == "Yes"]
# get the count of each profession in the agriculture sector
agriculture_data_scientists = agriculture["Data scientists"].value_counts()
print(agriculture_data_scientists)
agriculture_machine_learning_experts = agriculture["Machine learning experts"].value_counts()
print(agriculture_machine_learning_experts)
agriculture_software_engineers = agriculture["Software engineers"].value_counts()
print(agriculture_software_engineers)
agriculture_data_analysts = agriculture["Data analysts"].value_counts()
print(agriculture_data_analysts)
agriculture_statisticians = agriculture["Statisticians"].value_counts()
print(agriculture_statisticians)

# get the rows where the sector is Construction
construction = data[data["Construction"] == "Yes"]
# get the count of each profession in the construction sector
construction_data_scientists = construction["Data scientists"].value_counts()
print(construction_data_scientists)
construction_machine_learning_experts = construction["Machine learning experts"].value_counts()
print(construction_machine_learning_experts)
construction_software_engineers = construction["Software engineers"].value_counts()
print(construction_software_engineers)
construction_data_analysts = construction["Data analysts"].value_counts()
print(construction_data_analysts)
construction_statisticians = construction["Statisticians"].value_counts()
print(construction_statisticians) 

# get the rows where the sector is Creative Industries
creative = data[data["Creative Industries"] == "Yes"]
# get the count of each profession in the creative industries sector
creative_data_scientists = creative["Data scientists"].value_counts()
print(creative_data_scientists)
creative_machine_learning_experts = creative["Machine learning experts"].value_counts()
print(creative_machine_learning_experts)
creative_software_engineers = creative["Software engineers"].value_counts()
print(creative_software_engineers)
creative_data_analysts = creative["Data analysts"].value_counts()
print(creative_data_analysts)
creative_statisticians = creative["Statisticians"].value_counts()
print(creative_statisticians)

# get the rows where the sector is Transportation including logistics and warehousing
transportation = data[data["Transportation, including logistics and warehousing"] == "Yes"]
# get the count of each profession in the transportation sector
transportation_data_scientists = transportation["Data scientists"].value_counts()
print(transportation_data_scientists)
transportation_machine_learning_experts = transportation["Machine learning experts"].value_counts()
print(transportation_machine_learning_experts)
transportation_software_engineers = transportation["Software engineers"].value_counts()
print(transportation_software_engineers)
transportation_data_analysts = transportation["Data analysts"].value_counts()
print(transportation_data_analysts)
transportation_statisticians = transportation["Statisticians"].value_counts()
print(transportation_statisticians)

# get the rows where the sector is Other
other = data[data["Other (1)"] != "No"]
# get the count of each profession in the other sector
other_data_scientists = other["Data scientists"].value_counts()
print(other_data_scientists)
other_machine_learning_experts = other["Machine learning experts"].value_counts()
print(other_machine_learning_experts)
other_software_engineers = other["Software engineers"].value_counts()
print(other_software_engineers)
other_data_analysts = other["Data analysts"].value_counts()
print(other_data_analysts)
other_statisticians = other["Statisticians"].value_counts()
print(other_statisticians)

# For the column "Briefly outline what engagement you have had with data science and AI solutions in the past (for example, advice, training, specialised staff)", check the number of professionals for the given answer "Own data science and/or AI team internally"
own_data_science_team = data[data["Briefly outline what engagement you have had with data science and AI solutions in the past (for example, advice, training, specialised staff)"] == "Own data science and/or AI team internally"]

own_team_data_scientists = own_data_science_team["Data scientists"].value_counts()
print(own_team_data_scientists)
own_team_machine_learning_experts = own_data_science_team["Machine learning experts"].value_counts()
print(own_team_machine_learning_experts)
own_team_software_engineers = own_data_science_team["Software engineers"].value_counts()
print(own_team_software_engineers)
own_team_data_analysts = own_data_science_team["Data analysts"].value_counts()
print(own_team_data_analysts)
own_team_statisticians = own_data_science_team["Statisticians"].value_counts()
print(own_team_statisticians)

# Check all data if all professions are "Yes"
all_yes = data[(data["Data scientists"] == "Yes") & (data["Machine learning experts"] == "Yes") & (data["Software engineers"] == "Yes") & (data["Data analysts"] == "Yes") & (data["Statisticians"] == "Yes")]
print("Count of companies that have all professions:", all_yes.shape[0])
print(all_yes.to_string(index=False, header=False))


# What are the primary barriers that your company faces with regards to the adoption of AI? (choose up to 3)
# Categories: 
# - Company culture, approach to innovation
# - Data quality, data sharing and/or data management issues
# - Costs and complexity
# - Lack of strategy and clear objectives
# - Ethical, regulatory and compliance
# - Process deficiencies and storage
# - Other

# get the rows where the barrier is Company culture, approach to innovation
company_culture = data[data["Company culture, approach to innovation"] == "Yes"].shape[0]
data_quality = data[data["Data quality, data sharing and/or data management issues"] == "Yes"].shape[0]
costs = data[data["Costs and complexity"] == "Yes"].shape[0]
strategy = data[data["Lack of strategy and clear objectives"] == "Yes"].shape[0]
ethical = data[data["Ethical, regulatory and compliance"] == "Yes"].shape[0]
process = data[data["Process deficiencies and storage"] == "Yes"].shape[0]
other = data[data["Other (2)"] == "Yes"].shape[0]

# print all the counts in one line
print("Company culture:", company_culture, "Data quality:", data_quality, "Costs:", costs, "Strategy:", strategy, "Ethical:", ethical, "Process:", process, "Other:", other)



# Now conduct a secoral analysis of the barriers. For each sector, calculate the number of companies that face each barrier.
# get the rows where the sector is Agriculture and food processing
agriculture_company_culture = agriculture["Company culture, approach to innovation"].value_counts()
agriculture_data_quality = agriculture["Data quality, data sharing and/or data management issues"].value_counts()
agriculture_costs = agriculture["Costs and complexity"].value_counts()
agriculture_strategy = agriculture["Lack of strategy and clear objectives"].value_counts()
agriculture_ethical = agriculture["Ethical, regulatory and compliance"].value_counts()
agriculture_process = agriculture["Process deficiencies and storage"].value_counts()
agriculture_other = agriculture["Other (2)"].value_counts()

print("Company culture:", agriculture_company_culture)
print("Data quality:", agriculture_data_quality)
print("Costs:", agriculture_costs)
print("Strategy:", agriculture_strategy)
print("Ethical:", agriculture_ethical)
print("Process:", agriculture_process)
print("Other:", agriculture_other)

construction_company_culture = construction["Company culture, approach to innovation"].value_counts()
construction_data_quality = construction["Data quality, data sharing and/or data management issues"].value_counts()
construction_costs = construction["Costs and complexity"].value_counts()
construction_strategy = construction["Lack of strategy and clear objectives"].value_counts()
construction_ethical = construction["Ethical, regulatory and compliance"].value_counts()
construction_process = construction["Process deficiencies and storage"].value_counts()
construction_other = construction["Other (2)"].value_counts()

print("Company culture:", construction_company_culture)
print("Data quality:", construction_data_quality)
print("Costs:", construction_costs)
print("Strategy:", construction_strategy)
print("Ethical:", construction_ethical)
print("Process:", construction_process)
print("Other:", construction_other)

creative_company_culture = creative["Company culture, approach to innovation"].value_counts()
creative_data_quality = creative["Data quality, data sharing and/or data management issues"].value_counts()
creative_costs = creative["Costs and complexity"].value_counts()
creative_strategy = creative["Lack of strategy and clear objectives"].value_counts()
creative_ethical = creative["Ethical, regulatory and compliance"].value_counts()
creative_process = creative["Process deficiencies and storage"].value_counts()
creative_other = creative["Other (2)"].value_counts()

print("Company culture:", creative_company_culture)
print("Data quality:", creative_data_quality)
print("Costs:", creative_costs)
print("Strategy:", creative_strategy)
print("Ethical:", creative_ethical)
print("Process:", creative_process)
print("Other:", creative_other)

transportation_company_culture = transportation["Company culture, approach to innovation"].value_counts()
transportation_data_quality = transportation["Data quality, data sharing and/or data management issues"].value_counts()
transportation_costs = transportation["Costs and complexity"].value_counts()
transportation_strategy = transportation["Lack of strategy and clear objectives"].value_counts()
transportation_ethical = transportation["Ethical, regulatory and compliance"].value_counts()
transportation_process = transportation["Process deficiencies and storage"].value_counts()
transportation_other = transportation["Other (2)"].value_counts()

print("Company culture:", transportation_company_culture)
print("Data quality:", transportation_data_quality)
print("Costs:", transportation_costs)
print("Strategy:", transportation_strategy)
print("Ethical:", transportation_ethical)
print("Process:", transportation_process)
print("Other:", transportation_other)
