==> EoI_Process/analysis_notebooks/03-keyword-based.ipynb <==
import pandas as pd
from keybert import KeyBERT

data = pd.read_csv('../EoIs/Outputs/Combined/Inputs/merged_data.csv')

# get two columns related to business and AI challenge and combine it as a single text column
business = 'Business Challenge: ' + data['What is the business challenge that your organisation would most like to solve with data science and/or AI? (max 200 words)'] + '\n AI Challenge: ' + data['How might AI make your company more productive? (max 200 words)']

text = business[0]

kw_model = KeyBERT()
keywords = kw_model.extract_keywords(text)

keywords

# define some common words like ai and data to not bring them in the keywords
common_words = ['ai', 'data'] # --> then include it as stop_words in the extract_keywords method

kw_model.extract_keywords(text, keyphrase_ngram_range=(1, 2), stop_words=common_words)
