==> EoI_Process/analysis_notebooks/90-test-llm-inference.ipynb <==
import torch
from transformers import AutoModelForCausalLM, AutoTokenizer, BitsAndBytesConfig, Gemma3ForCausalLM

model_id = "google/gemma-3-1b-it"

#quantization_config = BitsAndBytesConfig(load_in_8bit=True)

model = Gemma3ForCausalLM.from_pretrained(
    model_id
).eval()

tokenizer = AutoTokenizer.from_pretrained(model_id)


**Other Sectors:**
- Healthcare: Including healthcare, digital health, medical device, etc.
- Environment, energy and sustainability: Water, urban development, climate resilience, wildlife conversation, etc.
- Information, communication  adn technology: Consultatation activities, cybersecurity, infrastructure, etc. 
- Finance: Fintech applications and financial inclusion
- Education: Higher education, Edtech, etc.
- Manufacturing: Chemical, aerospace, utilities, etc.
- Professional services: Including event management, food and drink, hospitality, 
- Legal: Patent, etc.
- Charities/NGOs

business_challenge = ""

messages = [{"role": "user", "content": "Summarise the following text that describes a business challenge of an SME in one sentence:" + business_challenge}]
input_text=tokenizer.apply_chat_template(messages, tokenize=False)
inputs = tokenizer.encode(input_text, return_tensors="pt")
outputs = model.generate(inputs, max_new_tokens=50, temperature=0.2, top_p=0.9, do_sample=True)

outputs


messages = [
    [
        {
            "role": "system",
            "content": [{"type": "text", "text": "You are a helpful assistant."},]
        },
        {
            "role": "user",
            "content": [{"type": "text", "text": "Write a poem on Hugging Face, the company"},]
        },
    ],
]
inputs = tokenizer.apply_chat_template(
    messages,
    add_generation_prompt=True,
    tokenize=True,
    return_dict=True,
    return_tensors="pt",
).to(model.device).to(torch.bfloat16)


with torch.inference_mode():
    outputs = model.generate(**inputs, max_new_tokens=64)

outputs = tokenizer.batch_decode(outputs)


outputs

print(tokenizer.decode(outputs[0]))
